Added a custom exception for Flashcard Pack

+++ SEProject/Services/FlashcardPackFileIOService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using SEProject.Models;
 using SEProject.EventArguments;
+using SEProject.Exceptions;
 
 namespace SEProject.Services;
 
@@ -65,6 +66,12 @@ public class FlashcardPackFileIOService : IFlashcardPackDataHandler
         var packToDelete = await _context.FlashcardPacks
         .Include(pack => pack.Flashcards)
         .FirstOrDefaultAsync(pack => pack.ID == ID);
+
+        if (packToDelete == null)
+        {
+            // Handle the case where the flashcardPack is null
+            throw new FlashcardPackNotFoundException($"FlashcardPack with ID {ID} not found");
+        }
 
         _context.FlashcardPacks.Remove(packToDelete!);
         _context.Flashcards.RemoveRange(packToDelete.Flashcards!);
